refactor(handlers): extract --map-model flag parsing to shared helper
Deduplicates the identical positional-flag scanning loop from install.ts
and update.ts into parsePositionalsWithFlags in shared.ts, resolving
the SonarCloud duplication gate failure.
